+++ easy-ui-react/src/TabPanels/TabPanels.tsx
@@ -1,19 +1,10 @@
-import { ListCollection } from "@react-stately/list";
-import { AriaLabelingProps, ItemElement, Node } from "@react-types/shared";
-import React, {
-  Key,
-  ReactElement,
-  ReactNode,
-  createContext,
-  useCallback,
-  useContext,
-  useEffect,
-  useMemo,
-  useState,
-} from "react";
-import { AriaTabPanelProps, useTab, useTabList, useTabPanel } from "react-aria";
-import { TabListState, useCollection, useTabListState } from "react-stately";
+import { AriaLabelingProps } from "@react-types/shared";
+import React, { Key, ReactNode, useMemo, useState } from "react";
+import { TabListState } from "react-stately";
 import { classNames } from "../utilities/css";
+import { TabPanelsContext } from "./context";
+import { TabPanelsPanels } from "./TabPanelsPanels";
+import { TabPanelsTabs } from "./TabPanelsTabs";
 
 import styles from "./TabPanels.module.scss";
 
@@ -53,47 +44,6 @@ type TabPanelsProps = AriaLabelingProps & {
   selectedKey?: Key | null;
 };
 
-type TabPanelsTabsProps = AriaLabelingProps & {
-  /**
-   * The tab items to display. Item keys should match the key of the
-   * corresponding <Item> within the <TabPanels.Panels> element.
-   */
-  children: ItemElement<object> | ItemElement<object>[];
-};
-
-type TabPanelsPanelsProps = {
-  /**
-   * The contents of each tab. Item keys should match the key of the
-   * corresponding <Item> within the <TabPanels.Tabs> element.
-   */
-  children: ReactElement | ReactElement[];
-};
-
-type TabProps = {
-  item: Node<object>;
-  state: TabListState<object>;
-};
-
-type TabPanelProps = AriaTabPanelProps & {
-  children: ReactNode;
-  state: TabListState<object>;
-};
-
-type TabPanelsContextType = {
-  tabListState: TabListState<object> | null;
-  setTabListState: (state: TabListState<object>) => void;
-};
-
-const TabPanelsContext = createContext<TabPanelsContextType | null>(null);
-
-const useTabPanels = () => {
-  const tabPanelsContext = useContext(TabPanelsContext);
-  if (!tabPanelsContext) {
-    throw new Error("useTabPanels must be used within a TabPanels");
-  }
-  return tabPanelsContext;
-};
-
 export function TabPanels(props: TabPanelsProps) {
   const { children } = props;
 
@@ -113,81 +63,5 @@ export function TabPanels(props: TabPanelsProps) {
   );
 }
 
-function TabPanelTabs(props: TabPanelsTabsProps) {
-  const { setTabListState } = useTabPanels();
-
-  const ref = React.useRef(null);
-
-  const state = useTabListState({ ...props, children: props.children });
-
-  const { tabListProps } = useTabList(props, state, ref);
-
-  useEffect(
-    () => {
-      setTabListState(state);
-    },
-    // state is too comprehensive of a dependency in and of itself. only change
-    // when specific requirements are met
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [state.disabledKeys, state.selectedItem, state.selectedKey, props.children],
-  );
-
-  return (
-    <div {...tabListProps} ref={ref}>
-      {[...state.collection].map((item) => (
-        <Tab key={item.key} item={item} state={state} />
-      ))}
-    </div>
-  );
-}
-
-function Tab({ item, state }: TabProps) {
-  const { key, rendered } = item;
-  const ref = React.useRef(null);
-  const { tabProps } = useTab({ key }, state, ref);
-  return (
-    <div {...tabProps} ref={ref}>
-      {rendered}
-    </div>
-  );
-}
-
-function TabPanelPanels(props: TabPanelsPanelsProps) {
-  const { tabListState } = useTabPanels();
-
-  const factory = useCallback(
-    (nodes: Iterable<Node<object>>) => new ListCollection(nodes),
-    [],
-  );
-
-  const collection = useCollection(props, factory, {
-    suppressTextValueWarning: true,
-  });
-
-  const selectedItem = tabListState
-    ? collection.getItem(tabListState.selectedKey)
-    : null;
-
-  if (!tabListState) {
-    return null;
-  }
-
-  return (
-    <TabPanel {...props} key={tabListState?.selectedKey} state={tabListState}>
-      {selectedItem && selectedItem.props.children}
-    </TabPanel>
-  );
-}
-
-function TabPanel({ state, ...props }: TabPanelProps) {
-  const ref = React.useRef(null);
-  const { tabPanelProps } = useTabPanel(props, state, ref);
-  return (
-    <div {...tabPanelProps} ref={ref}>
-      {props.children}
-    </div>
-  );
-}
-
-TabPanels.Tabs = TabPanelTabs;
-TabPanels.Panels = TabPanelPanels;
+TabPanels.Tabs = TabPanelsTabs;
+TabPanels.Panels = TabPanelsPanels;
